Fix cyclic deps
